contents()
UK Legislation - Table of Contents for an Act. contents() is a function with one argument url. When contents() is called, we pass along the URL corresponding to one Act's full data XML file. This URL is used inside the function for returning an instance of the Python object Parts that contains nested occurrences of the following objects: Chapters , cross-headings and Sections 

example links:  https://www.legislation.gov.uk/ukpga/2004/34/data.xml